% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{loss_function}
\alias{loss_function}
\alias{RSAVS_L2}
\alias{RSAVS_L1}
\alias{RSAVS_Huber}
\title{Built-in loss functions}
\usage{
RSAVS_L2(x, param)

RSAVS_L1(x, param)

RSAVS_Huber(x, param, derivative = FALSE)
}
\arguments{
\item{x}{input numeric vector}

\item{param}{parameters needed for the function, takes the form of 
numeric vector. Unused for L1 and L2.}

\item{derivative}{logical, whether the return is the loss value or the
derivative.}
}
\value{
Loss(or derivative) value at x.
}
\description{
These are built-in loss functions.
}
\examples{
RSAVS_L2(1 : 10)
RSAVS_L1(1 : 10)
RSAVS_Huber(seq(from = -3, to = 3, by = 0.1), param = 1.345)
RSAVS_Huber(seq(from = -3, to = 3, by = 0.1), param = 1.345, derivative = TRUE)
}
