% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty_function}
\alias{penalty_function}
\alias{penalty_lasso}
\alias{penalty_scad}
\alias{penalty_mcp}
\title{Built-in penalty functions}
\usage{
penalty_lasso(x, params, derivative = FALSE)

penalty_scad(x, params, derivative = FALSE)

penalty_mcp(x, params, derivative = FALSE)
}
\arguments{
\item{x}{input numeric vector}

\item{params}{numerical vector, parameters needed for the corresponding 
penalty function. Refer to the following sections for more details.}

\item{derivative}{boolen, whether the derivatives at \code{x} should be computed 
instead of the actual penalized value. Default to \code{FALSE}.}
}
\value{
A numerical vector storing the penalty function's value at \code{x} if
  \code{derivative = FALSE}. Otherwise the derivative of the penalty function at
  \code{x}
}
\description{
These are built-in penalty functions. They are implemented purely in R.
}
\section{Parameters for different penalties}{

  \itemize{
    \item For lasso, 
      \deqn{
        f(x, \lambda) = \lambda |x|
      }
      where \eqn{\lambda} = \code{param[1]}.
    \item For SCAD and MCP, \eqn{\lambda} = \code{param[1]} and \eqn{\gamma} = \code{param[2]}
  }
}

\examples{
x <- seq(from = -10, to = 10, by = 0.001)
lasso <- penalty_lasso(x, 1.0)
dlasso <- penalty_lasso(x, 1.0, TRUE)

scad <- penalty_scad(x, c(1.5, 3.7))
dscad <- penalty_scad(x, c(1.5, 3.7), TRUE)

mcp <- penalty_mcp(x, c(1.5, 2.7))
dmcp <- penalty_mcp(x, c(1.5, 2.7), TRUE)
  
matplot(x, cbind(lasso, scad, mcp), type = "l", ylab = "penalty")
legend(-10, 2.5, legend = c("lasso", "scad", "mcp"), col = 1 : 3, lty = 1 : 3)
matplot(x, cbind(dlasso, dscad, dmcp), type = "l", ylab = "derivative")
legend(-10, 1.5, legend = c("lasso", "scad", "mcp"), col = 1 : 3, lty = 1 : 3)
}
