\name{RSAVS-package}
\alias{RSAVS-package}
\alias{RSAVS}
\docType{package}
\title{
\packageTitle{RSAVS}
}
\description{
\packageDescription{RSAVS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RSAVS}
\packageIndices{RSAVS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RSAVS}

Maintainer: \packageMaintainer{RSAVS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# ~~ simple examples of the most important functions ~~
x <- sqrt(1 : 10)
print(x)
}
