% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSAVS.r
\docType{package}
\name{RSAVS}
\alias{RSAVS}
\title{RSAVS: A package for robust subgroup analysis and variable selection.}
\description{
The RSAVS can do subgroup analysis and variable selection simultaneously and 
it supports multiple choices of loss and penalty functions.
It implements the computation in a parallel manner.
}
\section{RSAVS functions}{

RSAVS_LargeN
}

