% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_Compute_BIC}
\alias{RSAVS_Compute_BIC}
\title{Compute Modified BIC}
\usage{
RSAVS_Compute_BIC(y_vec, x_mat, beta_vec, mu_vec, loss_type, loss_param, phi)
}
\arguments{
\item{y_vec}{numerical vector of response. \code{n = length(y_vec)} is the number of observations.}

\item{x_mat}{numerical matrix of covariates. \code{p = ncol(x_mat)} is the number of covariates.}

\item{beta_vec}{numerical vector of covariate effects}

\item{mu_vec}{numerical vector of subgroup effect(intercept term) for each observations}

\item{loss_type, loss_param}{type and parameters of the loss function
\itemize{
  \item \code{loss_type = "1"}: L-1 loss, no actual parameter is needed.
  \item \code{loss_type = "2"}: L-2 loss, no actual parameter is needed.
  \item \code{loss_type = "H"}, Huber loss, and \code{c = loss_param[0]}, where \code{c}
    is the parameter needed by huber loss. A popular choice is \code{c = 1.345}.
}}

\item{phi}{a positive constant.}
}
\value{
the mBIC value.
}
\description{
This function computes the modified BIC(mBIC), given a specific solution.
}
\details{
\deqn{BIC = log(1 / n * sum(loss(y - mu - x * beta)) + |S| * Phi ,}
where
  \itemize{
    \item mu is the intercept term of each observation. And the number of subgroups is
      \code{K = length(unique(mu_vec))}.
    \item beta is the covariate effect vector. And the number of active covariates is
      \code{Q = sum(beta_vec != 0)}.
    \item the loss function is determined by \code{loss_type} and \code{loss_param}.
    \item |S| is the complexity of the model and \eqn{|S| = K + Q}.
    \item Phi is a constant and \code{Phi = phi * log(log(n + p)) * log(n) / n}.
  }
}
