% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_Further_Improve}
\alias{RSAVS_Further_Improve}
\title{Post-selection estimation}
\usage{
RSAVS_Further_Improve(
  y_vec,
  x_mat,
  l_type = "L1",
  l_param = NULL,
  mu_vec,
  beta_vec
)
}
\arguments{
\item{y_vec}{numeric vector for response. 
\code{n = length(y_vec)} is the number of observations.}

\item{x_mat}{numeric matrix for covariates. Each row is for one observation.
\code{p = ncol(x_mat)} is the number of covariates.}

\item{l_type}{character, type of loss function.
\itemize{
  \item "L1": L-1 loss.
  \item "L2": L-2 loss.
  \item "Huber": Huber loss.
}
The default value is "1".}

\item{l_param}{vector of parameters needed by the loss function. 
For Huber loss, \code{c = l_param[1]} and a popular choice is c = 1.345.
The Default value is \code{NULL} since no additional parameters are
needed for the default L-1 loss.}

\item{mu_vec}{a length-n vector for subgroup effect.
The function uses this to determine the subgroup structure.}

\item{beta_vec}{a length-p vector for covariate effect.
The function uses this to find the active covariates.}
}
\value{
a list containing the improved mu(mu_vec) and beta(beta_vec) vector.
}
\description{
This function carries out the post-selection estimation given the model structure.
}
\details{
This function uses \code{mu_vec} to determine the subgroup structure matrix
  via \code{\link{RSAVS_Mu_to_Mat}}. The active covariates are those with non-zero
  \code{beta_vec} entries.
  
  Since this is the post-selection estimation. It's a regular estimate without
  any penalties. One should make sure the given model structure is identifiable.
}
\seealso{
\code{\link{RSAVS_Mu_to_Mat}} for getting the subgroup index matrix
  from the subgroup effect vector.
}
