% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RSAVS_Get_Lam_Max}
\alias{RSAVS_Get_Lam_Max}
\title{Compute lam_max}
\usage{
RSAVS_Get_Lam_Max(
  y_vec,
  x_mat,
  l_type = "L1",
  l_param = NULL,
  lam_ind = 1,
  const_abc = rep(1, 3),
  eps = 10^(-6)
)
}
\arguments{
\item{y_vec}{numeric vector, the response}

\item{x_mat}{numeric matrix, the covariate matrix}

\item{l_type}{character string, type of loss function.
\itemize{
  \item "L1": l-1 loss(absolute value loss)
  \item "L2": l-2 loss(squared error loss)
  \item "Huber": Huber loss. Its parameter is given in l_param.
}
Default value is "L1".}

\item{l_param}{numeric vector containing necessary parameters of the corresponding loss function. 
The default value is \code{NULL}.}

\item{lam_ind}{integer, indicating computation for \code{mu} or \code{beta}.
\itemize{
  \item 1: compute \code{lam_max} for \code{mu}.
  \item 2: compute \code{lam_max} for \code{beta}.
}}

\item{const_abc}{a length-3 numeric vector, providing the scalars to adjust weight
of regression function, penalty for subgroup identification and penalty for 
variable selection in the overall objective function. Defaults to \code{c(1, 1, 1)}.}

\item{eps}{a samll safe guard constant}
}
\value{
\code{lam_max}, a numerical variable
}
\description{
This function computes the values of \code{lam1_max} or \code{lam2_max} which 
  can shrink same type of covariates to the same value. 
  
For \code{beta}, they will be shrinked to 0. For \code{mu}, the final value 
is determined by the type of loss function. Please refer to \code{\link{RSAVS_Get_Psi}}
for more details. Basically, we have
\deqn{
  \mu_0 = \mathrm{argmin}_{\mu} \sum_i\rho(y_i - \mu)
}
and 
\deqn{
  psi\_vec = (\psi(y_1 - \mu_0), \cdots, \psi(y_n - \mu_0))
}
where \eqn{\psi = \partial \rho}.
}
\note{
For convex penalty like lasso, this \code{lam_max} can guarantee the global
  condition for shrinking the corresponding variables to the same value. For non-convex
  penalties such as SCAD and MCP, this is just a local condition.
}
\seealso{
\code{\link{RSAVS_Get_Psi}}.
}
