% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RSAVS_Get_Psi}
\alias{RSAVS_Get_Psi}
\title{Compute psi_vec}
\usage{
RSAVS_Get_Psi(y_vec, l_type = "L1", l_param = NULL)
}
\arguments{
\item{y_vec}{numeric vector}

\item{l_type}{character string, type of loss function.
\itemize{
  \item "L1": l-1 loss(absolute value loss)
  \item "L2": l-2 loss(squared error loss)
  \item "Huber": Huber loss. Its parameter is given in l_param.
}
Default value is "L1".}

\item{l_param}{numeric vector containing necessary parameters of the corresponding loss function. 
The default value is \code{NULL}.}
}
\value{
the \code{psi_vec}
}
\description{
This function computes the psi_vec needed by the algorithm.
}
\details{
In the algorithm, when all \eqn{\mu} are shrinked to the same value, we have
\deqn{
  \mu_0 = \mathrm{argmin}_{\mu} \sum_i\rho(y_i - \mu)
}
and 
\deqn{
  psi\_vec = (\psi(y_1 - \mu_0), \cdots, \psi(y_n - \mu_0))
}
where \eqn{\psi = \partial \rho}.
}
\note{
there is a SIGN different between this \code{psi_vec} and 
  \eqn{\partial \rho} w.r.t \eqn{\mu}

for non-differential loss like L1 at 0, the derivative at 0+ is returned.
}
