% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_Mu_to_Mat}
\alias{RSAVS_Mu_to_Mat}
\title{Generate the intercept term matrix according to mu_vec.}
\usage{
RSAVS_Mu_to_Mat(mu_vec)
}
\arguments{
\item{mu_vec:}{a length-n vector of the subgroup effect.}
}
\value{
res: a (n * p) matrix. Each row is for one observation and res[i, j] = 1 if i \eqn{/in} group_j and res[i, j] = 0 o.w.
}
\description{
This function generate the group id matrix according to the subgroup effect vector mu_vec.
}
\details{
\itemize{
\item n = length(mu_vec) is the number of observations.
\item p = length(unique(mu_vec)) is the number of subgroups.
}
}
\section{Note}{
 
In the result, 1st observation of mu_vec will always be in the 1st subgroup.
}

