% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_RI}
\alias{RSAVS_RI}
\title{Rand Index}
\usage{
RSAVS_RI(mu_est, mu_target, detail = FALSE)
}
\arguments{
\item{mu_est}{numerical vector of the estimated subgroup effect.
\code{n = length(mu_est)}.}

\item{mu_target}{numerical vector of the target(real) subgroup effect.
\code{n = length(mu_target)}.}

\item{detail}{whether or not should the function computes the details(TP, TN, FP, FN) 
of these 2 identifications(\code{mu_est} and \code{mu_target}).
Default value is \code{FALSE}.}
}
\value{
numerical value of Rand Index if \code{detail = FALSE}. 
  Otherwise a list containing \code{RI, TP, TN, FP, FN}.
}
\description{
This function computes the Rand Index(RI) of an estimated mu vector(mu_est) 
against the target mu vector(mu_target).
}
\section{Note}{

  \code{mu_est} and \code{mu_target} must have the same length.
}

\section{Details}{

  For a subgroup identification of \code{n} observations, 
  there are \eqn{n * (n - 1) / 2} pairs of different observations.
  The Rand Index describes how these two subgroup identifications are
  based on these observation pairs and it is given by
  \deqn{RI = (#TP + #TN) / (#TP + #TN + #FP + #FN),}
  where for any two observation \code{i} and \code{j},
  \itemize{
    \item TP(True Positive): 
      these two observations belong to the same subgroup in both these two identification.
    \item TN(True Negative): 
      these two observations belong to different subgroups in both these two identification.
    \item FP(False Positive): 
      these two observations belong to the same subgropu in \code{mu_est}.
      But they belong to different subgroups in \code{mu_target}.
    \item FN(False Negative): 
      these two observations belong to different subgroups in \code{mu_est}.
      But they belong to the same subgroup in \code{mu_target}.   
  }
  Therefore RI ranges from 0 to 1 and RI = 1 if and only if these two identifications are identical.
  
  Note: if \code{detail = TRUE}, the function also returns the number of TP, TN, FP and FN.
}

\examples{
n <- 50    # number of observations
# k subgroup effect, centered at 0
group_center <- seq(from = 0, to = 2 * (k - 1), by = 2) - (k - 1)
# subgroup effect vector    
alpha1 <- sample(group_center, size = n, replace = T) 
# another subgroup effect vector
alpha2 <- sample(group_center, size = n, replace = T)
RSAVS_RI(alpha1, alpha2)
RSAVS_RI(alpha1, alpha2, detail = TRUE)
}
