% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{RSAVS_Softthresh}
\alias{RSAVS_Softthresh}
\title{Soft thresh hold}
\usage{
RSAVS_Softthresh(invec, thresh)
}
\arguments{
\item{invec}{numerical vector, the original input value}

\item{thresh}{non-negative scalar, the threshholding point}
}
\value{
\code{outvec}. For each input entry \code{invec[i]}, the output value
  \code{outvec[i]} is
  \itemize{
    \item \code{outvec[i] = invec[i] - thresh}: if \code{invec[i] > thresh}
    \item \code{outvec[i] = invec[i] + thresh}: if \code{invec[i] < -thresh}
    \item \code{outvec[i] = 0}: if otherwise
  }
}
\description{
This function performs the so-called `soft threshholding` on the input 
vector(\code{invec}) based on the given threshholding value(\code{thresh})
}
\examples{
RSAVS_Softthresh(seq(-2, 2, by = 0.1), thresh = 1)
}
