% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{RSAVS_UpdateSW}
\alias{RSAVS_UpdateSW}
\alias{RSAVS_UpdateSW_Identity}
\alias{RSAVS_UpdateSW_Lasso}
\alias{RSAVS_UpdateSW_SCAD}
\alias{RSAVS_UpdateSW_MCP}
\title{Update S and W}
\usage{
RSAVS_UpdateSW_Identity(invec, param, const_r, const_bc)

RSAVS_UpdateSW_Lasso(invec, param, const_r, const_bc)

RSAVS_UpdateSW_SCAD(invec, param, const_r, const_bc)

RSAVS_UpdateSW_MCP(invec, param, const_r, const_bc)
}
\arguments{
\item{invec}{numerical vector, the orignal \code{s_vec} or \code{w_vec}.}

\item{param}{numerical vector, parameters needed for the penalty function.}

\item{const_r}{numerical scalar, parameter needed in the quadratic term in the 
augmented part.}

\item{const_bc}{numerical scalar, parameter represents the weight of corresponding
penalty part in the objective function}
}
\description{
These functions perform the updating step of \code{s_vec} and \code{w_vec} in 
  the ADMM algorithm for different types of penalty functions.
  
The updating steps of \code{s_vec} and \code{w_vec} share the same pattern in 
  the ADMM algorithm. These functions correspond to different penalty types
  (\code{p1_type} and \code{p2_type}) provided from the original solver function.
  \itemize{
    \item \code{UpdateSW_Identity}: identity update, this is needed when there
      is no shrinkage applied, possibly since the corresponding \code{lam} is 
      set to 0.
    \item \code{UpdateSW_Lasso}: Lasso update
    \item \code{UpdateSW_SCAD}: SCAD update
    \item \code{UpdateSW_MCP}: MCP update
  }
}
\note{
Please refer to the vignette about the algorithm detail design to find out
  more about how these parameters are defined.
}
\seealso{
\code{\link{RSAVS_Path_PureR}}, \code{\link{RSAVS_Solver_PureR}}
}
