% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_functions.R
\name{RSAVS_UpdateZ}
\alias{RSAVS_UpdateZ}
\alias{RSAVS_UpdateZ_L1}
\alias{RSAVS_UpdateZ_L2}
\alias{RSAVS_UpdateZ_Huber}
\title{Update Z}
\usage{
RSAVS_UpdateZ_L1(invec, param, r1, const_a)

RSAVS_UpdateZ_L2(invec, param, r1, const_a)

RSAVS_UpdateZ_Huber(invec, param, r1, const_a)
}
\arguments{
\item{invec}{numerical vector, the orignal \code{z_vec}}

\item{param}{numerical vector, parameters needed for the loss function. For "L1"
and "L2" loss, this will be ignored}

\item{r1}{numerical scalar, parameter needed in the quadratic term in the augmented
part for \code{z_vec}.}

\item{const_a}{numerical scalar, parameter controls the weight of regression part 
in the orignal objective function}
}
\description{
Update \code{z_vec} in the ADMM algorithm
}
\details{
These functions correspond to different types of loss function defined in the
  original problem.
\itemize{
  \item \code{RSAVS_UpdateZ_L1} corresponds to the "L1" loss, which means \code{l_type = "L1"}.
  \item \code{RSAVS_UpdateZ_L2} corresponds to the "L2" loss, which means \code{l_type = "L2"}.
  \item \code{RSAVS_UpdateZ_Huber} corresponds to the "Huber" loss, which means \code{l_type = "Huber"}.
}
}
\note{
When the loss type is "L2", it is possible to modify the algorithm to proceed 
  without \code{z_vec}. But currently for a more unified paradigm, this is not 
  implemented.

Please refer to the vignette about the algorithm detail design to find out
  more about how these parameters are defined.
}
\seealso{
\code{\link{RSAVS_Path_PureR}}, \code{\link{RSAVS_Solver_PureR}}
}
