% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_Determine_Mu}
\alias{RSAVS_Determine_Mu}
\title{Determine the subgroup effects}
\usage{
RSAVS_Determine_Mu(mu_vec, group_res)
}
\arguments{
\item{mu_vec}{numerical vector. \code{n = length(mu_vec)} is the number of observations.
Presumably it comes from the ADMM algorithm and is not a very good grouping result.}

\item{group_res}{a list, containing the grouping results.
Each element of group_res is a list containing the indecs from the same subgroup.
See \code{\link{RSAVS_S_to_Groups}} for the output structure.}
}
\value{
a length-n vector of subgroup effects. Observations belonging to the same
  subgroup should have the same effects.
}
\description{
This function determines the final mu vector given the grouping results.
}
\details{
Currently, the resulting subgrouping effects are the average value of \code{mu_vec}
  for those belonging to the same subgoup according to \code{group_res}.
  
  If \code{group_res} is missing, then a simple K-means is performed on \code{mu_vec}
  to get the estimated grouping result.
}
