% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_define_functions.r
\name{RSAVS_Generate_D_Matrix}
\alias{RSAVS_Generate_D_Matrix}
\title{Generate the pair-wise different matrix}
\usage{
RSAVS_Generate_D_Matrix(n, dense = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{dense}{logical, whether the return type should be in dense matrix or not}
}
\value{
a difference matrix with size \eqn{(n * (n - 1) / 2) \times n}
}
\description{
This function generate the pairwise difference matrix(D matrix in the paper)
By default it returns a sparse matrix(matrix.csr) from the package SparseM
}
